Refactor: SpeechException 추가, ErrorCode 작성
