+++ src/main/java/com/vowser/backend/application/service/speech/SpeechService.java
@@ -1,15 +1,15 @@
-package com.vowser.backend.application.service;
+package com.vowser.backend.application.service.speech;
 
 import com.google.cloud.speech.v2.*;
 import com.google.protobuf.ByteString;
-import com.vowser.backend.common.constants.ErrorMessages;
 import com.vowser.backend.common.constants.NetworkConstants;
 import com.vowser.backend.common.constants.SpeechConstants;
+import com.vowser.backend.common.exception.ErrorCode;
+import com.vowser.backend.common.exception.SpeechException;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
-import org.springframework.util.Assert;
 import org.springframework.web.multipart.MultipartFile;
 
 import java.io.IOException;
@@ -34,22 +34,29 @@ public class SpeechService {
     /**
      * Google Cloud Speech-to-Text를 사용하여 오디오 파일을 텍스트로 변환
      *
-     * @param audioFile 변환할 업로드된 오디오 파일
-     * @return 오디오로부터 변환된 텍스트
-     * @throws IllegalArgumentException 오디오 파일이 비어있거나 유효하지 않은 경우
-     * @throws RuntimeException 변환(인식) 과정이 실패한 경우
+     * @param audioFile 변환할 오디오 파일
+     * @return 인식된 텍스트(양끝 공백 제거)
+     * @throws SpeechException 입력이 유효하지 않거나 파일 읽기/인식 과정에서 오류가 발생한 경우
      */
     public String transcribe(MultipartFile audioFile) {
-        log.info("음성 파일 수신: {}, 크기: {} KB", 
-                audioFile.getOriginalFilename(), 
+
+        if (audioFile == null) {
+            throw new SpeechException(ErrorCode.INVALID_INPUT_VALUE, "audioFile is null");
+        }
+        if (audioFile.isEmpty()) {
+            throw new SpeechException(
+                    ErrorCode.SPEECH_EMPTY_AUDIO_FILE,
+                    "filename=" + audioFile.getOriginalFilename()
+            );
+        }
+        log.info("음성 파일 수신: {}, 크기: {} KB",
+                audioFile.getOriginalFilename(),
                 audioFile.getSize() / NetworkConstants.DataSize.BYTES_PER_KB);
-        
-        Assert.notNull(audioFile, "Audio file must not be null");
-        Assert.isTrue(!audioFile.isEmpty(), ErrorMessages.Speech.EMPTY_AUDIO_FILE);
+
 
         try {
             ByteString audioData = ByteString.copyFrom(audioFile.getBytes());
-            
+
             RecognitionConfig config = RecognitionConfig.newBuilder()
                     .addLanguageCodes(SpeechConstants.Language.KOREAN)
                     .setModel(SpeechConstants.Model.LONG)
@@ -66,24 +73,24 @@ public class SpeechService {
                     .build();
 
             log.info("Google Speech-to-Text API 호출 중... (recognizer: {})", recognizerPath);
-            
+
             RecognizeResponse response = speechClient.recognize(request);
 
             if (response.getResultsList().isEmpty()) {
-                throw new RuntimeException(ErrorMessages.Speech.NO_SPEECH_RECOGNIZED);
+                throw new SpeechException(ErrorCode.SPEECH_NO_SPEECH_RECOGNIZED);
             }
 
             String transcript = response.getResults(0).getAlternatives(0).getTranscript();
-            
+
             log.info("음성 인식 성공 - 변환된 텍스트: [{}]", transcript);
             return transcript.trim();
 
         } catch (IOException e) {
             log.error("음성 파일 읽기 실패: {}", audioFile.getOriginalFilename(), e);
-            throw new RuntimeException(ErrorMessages.Speech.CANNOT_READ_AUDIO_FILE + e.getMessage(), e);
+            throw new SpeechException(ErrorCode.SPEECH_CANNOT_READ_AUDIO_FILE);
         } catch (Exception e) {
             log.error("음성 인식 처리 실패: {}", audioFile.getOriginalFilename(), e);
-            throw new RuntimeException(ErrorMessages.Speech.SPEECH_RECOGNITION_FAILED + e.getMessage(), e);
+            throw new SpeechException(ErrorCode.SPEECH_RECOGNITION_FAILED);
         }
     }
 }
