+++ src/main/java/com/vowser/backend/common/exception/SpeechException.java
@@ -1,0 +1,14 @@
+package com.vowser.backend.common.exception;
+
+public class SpeechException extends BaseException {
+    public SpeechException(ErrorCode errorCode) {
+        super(errorCode);
+    }
+
+    public SpeechException(
+            ErrorCode errorCode,
+            String detail
+    ) {
+        super(errorCode, detail);
+    }
+}
